For M1 support, Stop building {fmt} as header-only
We were seeing the following when doing a Release build on an M1 mac:

    dlopen([snip]/_qwerty_harness.abi3.so, 0x0002): symbol not found in flat namespace '__ZN3fmt2v87vformatENS0_17basic_string_viewIcEENS0_17basic_format_argsINS0_20basic_format_contextINS0_8appenderEcEEEE'

When you demangle that symbol, you get

    _fmt::v8::vformat(fmt::v8::basic_string_view<char>, fmt::v8::basic_format_args<fmt::v8::basic_format_context<fmt::v8::appender, char> >)

This issue was previously reported for fmt [1][2], and it was
suggested[2] to disable {fmt} being built as header-only.
This indeed seems to resolve the issue; the only downside is that
consumers of tweedledum must link with libfmt.a as well as
libtweedledum.a. This is acceptable for Qwerty.

[1]: https://github.com/fmtlib/fmt/issues/3505
[2]: https://github.com/gabime/spdlog/issues/2773